Moves MetaData to types
