+++ sdk/go/init.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/base"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/logger"
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/rpc"
+	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/types"
 	initPkg "github.com/TrueBlocks/trueblocks-core/src/apps/chifra/sdk"
 	// EXISTING_CODE
 )
@@ -73,7 +73,7 @@ func GetInitOptions(args []string) (*InitOptions, error) {
 	return &opts, nil
 }
 
-func (opts *InitOptions) Query() ([]bool, *rpc.MetaData, error) {
+func (opts *InitOptions) Query() ([]bool, *types.MetaData, error) {
 	buffer := bytes.Buffer{}
 	if err := opts.Init(&buffer); err != nil {
 		logger.Fatal(err)
