+++ sdk/go/result.go
@@ -1,13 +1,8 @@
 package sdk
 
-import "github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/rpc"
-
-// type abisResult struct {
-// 	Data []bool       `json:"data"`
-// 	Meta rpc.MetaData `json:"meta"`
-// }
+import "github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/types"
 
 type Result[T any] struct {
-	Data []T          `json:"data"`
-	Meta rpc.MetaData `json:"meta"`
+	Data []T            `json:"data"`
+	Meta types.MetaData `json:"meta"`
 }
