+++ sdk/go/slurp.go
@@ -18,7 +18,6 @@ import (
 	"strings"
 
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/logger"
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/rpc"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/types"
 	slurp "github.com/TrueBlocks/trueblocks-core/src/apps/chifra/sdk"
 	// EXISTING_CODE
@@ -97,7 +96,7 @@ func GetSlurpOptions(args []string) (*SlurpOptions, error) {
 	return &opts, nil
 }
 
-func (opts *SlurpOptions) Query() ([]types.SimpleSlurp, *rpc.MetaData, error) {
+func (opts *SlurpOptions) Query() ([]types.SimpleSlurp, *types.MetaData, error) {
 	buffer := bytes.Buffer{}
 	if err := opts.Slurp(&buffer); err != nil {
 		logger.Fatal(err)
