+++ src/apps/chifra/pkg/notify/notification.go
@@ -2,6 +2,8 @@ package notify
 
 import "github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/rpc"
 
+// TODO: Remove rpc.MetaData and use types.MetaData throughout
+
 type Message string
 
 type Notification[T NotificationPayload] struct {
