+++ src/apps/chifra/pkg/rpc/get_meta.go
@@ -6,23 +6,13 @@ package rpc
 
 import (
 	"context"
-	"encoding/json"
 
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/base"
+	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/types"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/utils"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/walk"
 )
 
-type MetaData struct {
-	Latest    uint64 `json:"client"`
-	Finalized uint64 `json:"finalized"`
-	Staging   uint64 `json:"staging"`
-	Ripe      uint64 `json:"ripe"`
-	Unripe    uint64 `json:"unripe"`
-	ChainId   uint64 `json:"chainId,omitempty"`
-	NetworkId uint64 `json:"networkId,omitempty"`
-	Chain     string `json:"chain,omitempty"`
-}
+// TODO: use types.MetaData throughout
 
 func (conn *Connection) GetMetaData(testmode bool) (*MetaData, error) {
 	chainId, networkId, err := conn.GetClientIDs()
@@ -86,27 +76,6 @@ func (conn *Connection) GetMetaData(testmode bool) (*MetaData, error) {
 	return &meta, nil
 }
 
-func (m *MetaData) String() string {
-	ret, _ := json.MarshalIndent(m, "", "  ")
-	return string(ret)
-}
+// TODO: Remove rpc.MetaData and use types.MetaData throughout
 
-// Highest returns the height of the index (i.e., max between the finalized, staging, and ripe indexes).
-func (m *MetaData) IndexHeight() base.Blknum {
-	return utils.Max(m.Finalized, utils.Max(m.Staging, m.Ripe))
-}
-
-// NextIndexHeight returns the block after the height of the index.
-func (m *MetaData) NextIndexHeight() base.Blknum {
-	return m.IndexHeight() + 1
-}
-
-// ChainHeight returns the block after the height of the index.
-func (m *MetaData) ChainHeight() base.Blknum {
-	return m.Latest
-}
-
-// StageHieght returns the highest block that's been staged
-func (m *MetaData) StageHeight() base.Blknum {
-	return m.Staging
-}
+type MetaData = types.MetaData
