+++ src/dev_tools/goMaker/templates/classDefinitions/status.toml
@@ -4,4 +4,4 @@
     doc_descr = "status-related data about the TrueBlocks system including the server and local binary caches"
     doc_route = "403-status"
     produced_by = "status"
-    go_output = "src/apps/chifra/internal/status"
+    go_output = "src/apps/chifra/pkg/types"
